+++ packages/data-broker/src/services/dhis/DhisTranslator.js
@@ -7,13 +7,13 @@ import keyBy from 'lodash.keyby';
 
 import { translateElementKeysInEventAnalytics } from '@tupaia/dhis-api';
 import { reduceToDictionary } from '@tupaia/utils';
-import { InboundAggregateDataTranslator } from './InboundAggregateDataTranslator';
+import { InboundAnalyticsTranslator } from './InboundAnalyticsTranslator';
 import { parseValueForDhis } from './parseValueForDhis';
 
 export class DhisTranslator {
   constructor(models) {
     this.models = models;
-    this.inboundAggregateDataTranslator = new InboundAggregateDataTranslator();
+    this.inboundAnalyticsTranslator = new InboundAnalyticsTranslator();
     this.dataElementsByCode = {};
   }
 
@@ -136,8 +136,8 @@ export class DhisTranslator {
     return { ...restOfEvent, dataValues: translatedDataValues };
   }
 
-  translateInboundAggregateData = (response, dataSources) => {
-    return this.inboundAggregateDataTranslator.translate(response, dataSources);
+  translateInboundAnalytics = (response, dataSources) => {
+    return this.inboundAnalyticsTranslator.translate(response, dataSources);
   };
 
   translateInboundEventDataValues = (dataValues, dataElementToSourceCode) => {
@@ -204,15 +204,11 @@ export class DhisTranslator {
   };
 
   translateInboundIndicators = (indicators, dataSources) => {
-    const dataElementToSourceCode = reduceToDictionary(dataSources, 'dataElementCode', 'code');
     const indicatorIdToSourceCode = reduceToDictionary(dataSources, d => d.config?.dhisId, 'code');
 
     return indicators.map(({ code, ...restOfIndicators }) => {
       const translatedIndicators = { ...restOfIndicators };
       const { id } = translatedIndicators;
-      if (code) {
-        translatedIndicators.code = dataElementToSourceCode[code];
-      }
       if (!translatedIndicators.code && id) {
         translatedIndicators.code = indicatorIdToSourceCode[id];
       }
