+++ packages/data-broker/src/services/dhis/InboundAnalyticsTranslator.js
@@ -15,7 +15,7 @@ const DIMENSIONS = {
 const createDataElementKey = (dataElement, categoryOptionCombo) =>
   categoryOptionCombo ? `${dataElement}:${categoryOptionCombo}` : dataElement;
 
-export class InboundAggregateDataTranslator {
+export class InboundAnalyticsTranslator {
   findDimensionRowIndex = dimension => this.headers.findIndex(({ name }) => name === dimension);
 
   getDimensionIds = dimension => this.metadata.dimensions[dimension] || [];
@@ -103,12 +103,7 @@ export class InboundAggregateDataTranslator {
       this.dataSources,
       ({ dataElementCode }) => dataElementCode,
     );
-    // We can't always guarantee that dhis data elements can have codes (eg: Laos DHIS).
-    // So we should also have mapping between dhisId and data sources.
-    const dhisIdToDataSources = groupBy(
-      this.dataSources.filter(d => d.config.dhisId !== undefined),
-      ({ config }) => config.dhisId,
-    );
+
     const translatedItems = {};
     const translatedDx = [];
     Object.values(items).forEach(item => {
@@ -127,7 +122,6 @@ export class InboundAggregateDataTranslator {
       const { items: newItems, dx: newDx } = this.getTranslatedDataElementItemAndDx(
         item,
         dataElementCodeToDataSources,
-        dhisIdToDataSources,
         coCombosByCode,
       );
       Object.assign(translatedItems, newItems);
@@ -137,25 +131,18 @@ export class InboundAggregateDataTranslator {
     return { items: translatedItems, dx: translatedDx };
   }
 
-  getTranslatedDataElementItemAndDx(
-    item,
-    dataElementCodeToDataSources,
-    dhisIdToDataSources,
-    coCombosByCode,
-  ) {
-    const { uid, code: dhisCode, name } = item;
+  getTranslatedDataElementItemAndDx(item, dataElementCodeToDataSources, coCombosByCode) {
+    const { uid, code, name } = item;
 
     const translatedItems = {};
     const translatedDx = [];
-    // if dhis data element have no codes, find it by uids instead
-    const dataSourcesForElement =
-      dataElementCodeToDataSources[dhisCode] || dhisIdToDataSources[uid];
+    const dataSourcesForElement = dataElementCodeToDataSources[code];
     dataSourcesForElement.forEach(dataSource => {
       const coComboCode = dataSource.config.categoryOptionCombo;
       const { name: coComboName } = coCombosByCode[coComboCode] || {};
       const translatedId = createDataElementKey(uid, coComboCode);
       translatedDx.push(translatedId);
-      const dataElementKey = createDataElementKey(dhisCode, coComboCode);
+      const dataElementKey = createDataElementKey(code, coComboCode);
       translatedItems[translatedId] = {
         ...item,
         uid: translatedId,
