+++ packages/utils/src/period/index.js
@@ -8,6 +8,7 @@ export {
   convertDateRangeToPeriodString,
 } from './convertDateRangeToPeriods';
 export { convertPeriodStringToDateRange } from './convertPeriodStringToDateRange';
+export { convertDateRangeToPeriodQueryString } from './convertDateRangeToPeriodQueryString';
 export { getMostRecentPeriod, getMostAncientPeriod } from './periodExtremes';
 export * from './getDefaultPeriod';
 export * from './period';
