+++ packages/web-config-server/src/tests/apiV1/dataBuilders/generic/table/simpleTableOfEvents.test.js
@@ -75,7 +75,6 @@ fetchAnalytics
       dataServices,
       entityAggregation: undefined,
       dataSourceEntityFilter: undefined,
-      dataPeriodType: undefined,
       programCodes: ['WSRS'],
     },
     query,
